ManyToMany version

+++ UseCases.API/Deliveries/Commands/DeleteDelivery.cs
@@ -17,6 +17,7 @@ namespace UseCases.API.Deliveries
             public CommandHandler(DataContext context) => _context = context;
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
+                if (_context.Deliveries == null) return Unit.Value;
                 Delivery? Delivery = await _context.Deliveries.FindAsync(new object?[] { request.Id }, cancellationToken: cancellationToken);
                 if (Delivery == null) return Unit.Value;
                 _context.Deliveries.Remove(Delivery);
