+++ UseCases.API/OrderItems/DeleteOrderItem.cs
@@ -15,7 +15,11 @@ namespace UseCases.API.OrderItems
             public CommandHandler(DataContext context) => _context = context;
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
-                OrderItem orderItem = await _context.OrderItems.FindAsync(new object?[] { request.Id }, cancellationToken: cancellationToken);
+                if (_context.OrderItems == null)
+                {
+                    return Unit.Value;
+                }
+                OrderItem? orderItem = await _context.OrderItems.FindAsync(new object?[] { request.Id }, cancellationToken: cancellationToken);
                 if (orderItem == null) return Unit.Value;
                 _context.OrderItems.Remove(orderItem);
                 await _context.SaveChangesAsync(cancellationToken);
