+++ UseCases.API/Products/DeleteProduct.cs
@@ -15,6 +15,10 @@ namespace UseCases.API.Products
             public CommandHandler(DataContext context) => _context = context;
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
+                if (_context.Products==null)
+                {
+                    return Unit.Value;
+                }
                 Product? product = await _context.Products.FindAsync(new object?[] { request.Id }, cancellationToken: cancellationToken);
                 if (product == null) return Unit.Value;
                 _context.Products.Remove(product);
